+++ docs/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ----
 
+## v2.1.2 (2022-01-28)
+
+- Added so that `Parse` function can except not only past date but also future date and return correct result. Closes #23;
+- Added section `12 Features` to the `README.md`;
+
+----
+
 ## v2.1.1 (2022-01-28)
 
 - Added option `justNow` that prints `Just now` if time is within 60 minutes;